+++ src/AbstractUnitTestCase.ts
@@ -1,10 +1,11 @@
 import {Assert} from "./Assert";
 import {ObjectBuilder} from "./Mock";
+import {ExpectedExceptionContainer} from "./Test/ExpectedExceptionContainer";
 
 export class AbstractUnitTestCase {
 
     protected numberOfAsserts: number = 0;
-    protected expectExceptionClass;
+    protected expectExceptionContainer = null;
 
     protected assertTrue(expr: boolean, msg?: string): void {
         this.numberOfAsserts++;
@@ -58,23 +59,25 @@ export class AbstractUnitTestCase {
 
     protected expectException(exceptionClass, msg?: string): void {
         this.numberOfAsserts++;
-        this.expectExceptionClass = exceptionClass;
+        this.expectExceptionContainer = new ExpectedExceptionContainer(exceptionClass, msg);
     }
 
     public hasExpectedException(): boolean {
-        return new Boolean(this.expectExceptionClass) == true;
+        return this.expectExceptionContainer !== null;
     }
 
-    public pullExpectedException() {
-        let expectedException = this.expectExceptionClass;
-        this.expectExceptionClass = null;
+    /**
+     * @internal
+     */
+    public pullExpectedException(): ExpectedExceptionContainer {
+        let expectedException = this.expectExceptionContainer;
+        this.expectExceptionContainer = null;
         return expectedException;
     }
 
     public getNumberOfAssertions(): number {
         return this.numberOfAsserts;
     }
-
 
     /**
      * This method will be called once before class cases be executed
@@ -100,12 +103,7 @@ export class AbstractUnitTestCase {
     public tearDownAfterClass(): void {
     }
 
-    /**
-     *
-     * @param className
-     * @returns {T}
-     */
-    protected getMockBuilder(className) {
+    protected getMockBuilder(className): ObjectBuilder {
         return new ObjectBuilder(className);
     }