+++ tests/Assert/AssertTest.ts
@@ -205,14 +205,8 @@ export class AssertTest extends AbstractUnitTestCase {
     }
 
     public testAssertContainsInvalidArgument() {
-        let expectedException;
-        try {
-            this.assertContains("needle", 100500);
-        } catch (e) {
-            expectedException = e;
-        } finally {
-            this.assertInstanceOf(InvalidArgumentException, expectedException);
-        }
+        this.expectException(InvalidArgumentException);
+        this.assertContains("needle", 100500);
     }
 
     public testAssertInstanceOf() {
@@ -233,11 +227,6 @@ export class AssertTest extends AbstractUnitTestCase {
             let message = expectedException.getMessage();
             this.assertEquals("Failed asserting that object instance of AbstractUnitTestCase. ", message);
         }
-    }
-
-    public testExpectException() {
-        this.expectException(Error);
-        throw new Error();
     }
 
 }