+++ CHANGELOG.md
@@ -4,10 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
-### Added 
-- AbstractUnitTestCase.expectedException() method. TODO: use container for compact exception with user defined message.  
-
 See [ROADMAP](ROADMAP.md).
+
+## v0.12.0 - 2016-10-18
+### Added 
+- AbstractUnitTestCase.expectedException() method.    
 
 ## v0.11.1 - 2016-10-15
 ### Changed